fix: fix fatal error on model save if object cache is disabled

+++ bedita-app/models/behaviors/cacheable.php
@@ -199,6 +199,10 @@ class CacheableBehavior extends ModelBehavior {
             $this->addObjectsToClean($model, $this->getObjectsToCleanById($model, $objectId));
         }
 
+        if (!$this->on) {
+            return;
+        }
+
         foreach ($this->objectsToClean as $id) {
             $this->BeObjectCache->delete($id);
         }
